Renamed smclibcpp to smclib
