+++ bond_core/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="2">
+<!-- <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?> -->
+<package format="3">
   <name>bond_core</name>
   <version>2.0.0</version>
   <description>
@@ -21,7 +21,7 @@
   <exec_depend>bond</exec_depend>
   <exec_depend>bondcpp</exec_depend>
   <!-- <exec_depend>bondpy</exec_depend>-->
-  <exec_depend>smclibcpp</exec_depend>
+  <exec_depend>smclib</exec_depend>
   <exec_depend>test_bond</exec_depend>
   <export>
     <build_type>ament_cmake</build_type>
