+++ bondcpp/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="2">
+<!-- <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?> -->
+<package format="3">
   <name>bondcpp</name>
   <version>2.0.0</version>
   <description>
@@ -27,16 +27,15 @@
   <build_depend>rcl_interfaces</build_depend>
   <build_depend>rcl</build_depend>
   <build_depend>bond</build_depend>
-  <build_depend>boost</build_depend>
   <build_depend>rclcpp</build_depend>
-  <build_depend>smclibcpp</build_depend>
+  <build_depend>smclib</build_depend>
   <build_depend>uuid</build_depend>
   
   <exec_depend>rclcpp</exec_depend>
   <exec_depend>rcutils</exec_depend>
   <exec_depend>rmw</exec_depend>
   <exec_depend>bond</exec_depend>
-  <exec_depend>smclibcpp</exec_depend>
+  <exec_depend>smclib</exec_depend>
   <exec_depend>uuid</exec_depend>
   
   <test_depend>ament_lint_auto</test_depend>
