+++ smclib/CHANGELOG.rst
