+++ smclib/mainpage.dox
