+++ smclib/package.xml
@@ -1,5 +1,7 @@
-<package format="2">
-<name>smclibcpp</name>
+<?xml version="1.0"?>
+<!-- <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?> -->
+<package format="3">
+<name>smclib</name>
 <version>2.0.0</version>
 <description>
     The State Machine Compiler (SMC) from http://smc.sourceforge.net/
